pixel complex added
Produces a new image that only contains the values 0 or 255 in each
position. The pixel is assigned a value of 255 off the average value in
the adjacent pixels in the original image (including diagonals) is
lower than the pixel's value itself in the original image. Otherwise,
the value is 0. Prints the result of the edge detection, in a grid
format.